sched: address split follow-up comments and add sched ctests
