+++ kernel/include/sched/sched_ai_guard.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_AI_GUARD_H
+#define BHARAT_SCHED_AI_GUARD_H
+#include "sched/sched.h"
+#endif
