+++ kernel/include/sched/sched_ai_ingest.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_AI_INGEST_H
+#define BHARAT_SCHED_AI_INGEST_H
+#include "sched/sched.h"
+#endif
