+++ kernel/include/sched/sched_core.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_CORE_H
+#define BHARAT_SCHED_CORE_H
+#include "sched/sched.h"
+#endif
