+++ kernel/include/sched/sched_gp.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_GP_H
+#define BHARAT_SCHED_GP_H
+#include "sched/sched.h"
+#endif
