+++ kernel/include/sched/sched_migrate.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_MIGRATE_H
+#define BHARAT_SCHED_MIGRATE_H
+#include "sched/sched.h"
+#endif
