+++ kernel/include/sched/sched_pi.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_PI_H
+#define BHARAT_SCHED_PI_H
+#include "sched/sched.h"
+#endif
