+++ kernel/include/sched/sched_router.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_ROUTER_H
+#define BHARAT_SCHED_ROUTER_H
+#include "sched/sched.h"
+#endif
