+++ kernel/include/sched/sched_rt.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_RT_H
+#define BHARAT_SCHED_RT_H
+#include "sched/sched.h"
+#endif
