+++ kernel/include/sched/sched_runqueue.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_RUNQUEUE_H
+#define BHARAT_SCHED_RUNQUEUE_H
+#include "sched/sched.h"
+#endif
