+++ kernel/include/sched/sched_thread.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_THREAD_H
+#define BHARAT_SCHED_THREAD_H
+#include "sched/sched.h"
+#endif
