+++ kernel/include/sched/sched_wait.h
@@ -1,0 +1,4 @@
+#ifndef BHARAT_SCHED_WAIT_H
+#define BHARAT_SCHED_WAIT_H
+#include "sched/sched.h"
+#endif
