+++ kernel/src/sched/CMakeLists.txt
@@ -1,5 +1,16 @@
 add_library(k_sched OBJECT
     sched.c
+    sched_core.c
+    sched_runqueue.c
+    sched_thread.c
+    sched_wait.c
+    sched_rt.c
+    sched_gp.c
+    sched_migrate.c
+    sched_pi.c
+    sched_ai_ingest.c
+    sched_ai_guard.c
+    sched_router.c
     sched_deg.c
     algo_matrix.c
     ai_sched.c
@@ -9,3 +20,12 @@ target_sources(k_sched PRIVATE sched_test_support.c)
 
 target_include_directories(k_sched PRIVATE ../../include ../../include/generated ${CMAKE_SOURCE_DIR}/include ${CMAKE_SOURCE_DIR}/lib/include ${CMAKE_SOURCE_DIR}/boot/include)
 target_link_libraries(k_sched PRIVATE bharat_freestanding bharat_kernel_buildopts)
+
+add_test(
+    NAME sched.k_sched.build
+    COMMAND ${CMAKE_COMMAND} --build ${CMAKE_BINARY_DIR} --target k_sched
+)
+add_test(
+    NAME sched.k_test_support.build
+    COMMAND ${CMAKE_COMMAND} --build ${CMAKE_BINARY_DIR} --target k_test_support
+)
